car_classifier/models/backlight_detector/requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy
Pillow
pycocotools